gets the seperate tables to work correctly. still need to populate the optical table data correctly. but we're now much closer.
